transport: remove Open() from the transport interface
This abstracts out serial port/tcp socket management from transports,
allowing them to be used on any object implemeting a subset of the
net.Conn interface.

The end-goal is to allow the tcp transport to be used on both
client and server sockets.
Having the connection management logic in the client (and later on
in the server) rather than in transports also makes for clearer
code and interfaces.