server: add modbus/tcp over TLS support
This adds support for tunneling modbus TCP over TLS to the server
(sometimes called MBAPS or Modbus Security).

Connections are secured with TLSv1.2 or later and TLS mutual auth
(client + server authentication with TLS certs) is enforced, as
required in the Modbus Security/MBAPS specification
(MB-TCP-Security-v21_2018-07-24).

The Modbus Role value contained in the client certificate is exposed
to the request handler, allowing custom role-based autorisation schemes.