server: fix race condition
This fixes a race condition in the server code where a client
connection could be added to the active connection pool if accepted
while the server was being stopped.
The handler goroutine would keep on serving that connection instead
of being closed.

Credits to Luc Arnaud for spotting this bug.