client: allow reading up to 125 registers
This fixes a typo where the max quantity of registers that can be
read at once was set to 123.
While 123 is the maximum quantity of registers that can be written
at once, the limit for reading is 125 registers.

This also fixes a few logic errors when exceeding those limits
where the client wouldn't always return ErrUnexpectedParameters and
would sometimes proceed with the request.