Fix build on OpenBSD #60
modbus didn't build on OpenBSD because goburrow/serial doesn't
support it in the most recently released version. However it
does support it in a commit made subsequently.

Unfortunately the goburrow/serial project has not seen activity
since this commit, so the best we can do it pin it by commit ID
rather than a tag.